+++ world/.PAUL.wbproj
@@ -1,5 +1,5 @@
 Webots Project File version R2021a
-perspectives: 000000ff00000000fd000000030000000000000087000002fdfc0100000002fb0000001e00480074006d006c0052006f0062006f007400570069006e0064006f00770000000000000000870000000000000000fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000ffffffff0000008700ffffff00000001000001a0000002fdfc0200000001fb0000001400540065007800740045006400690074006f00720000000000000002fd0000004500ffffff00000003000005a0000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000005a00000000000000000000005a0000002ff00000001000000020000000100000008fc00000000
+perspectives: 000000ff00000000fd000000030000000000000087000002fdfc0100000002fb0000001e00480074006d006c0052006f0062006f007400570069006e0064006f00770000000000000000870000000000000000fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000ffffffff0000008700ffffff00000001000001a0000002fdfc0200000001fb0000001400540065007800740045006400690074006f00720000000000000002fd0000004500ffffff00000003000005a0000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000005a00000000000000000000005a00000030500000001000000020000000100000008fc00000000
 simulationViewPerspectives: 000000ff0000000100000002000001140000036e0100000002010000000100
 sceneTreePerspectives: 000000ff0000000100000002000000c0000000fa0100000002010000000200
 maximizedDockId: -1
