+++ worlds/.with UV module.wbproj
@@ -1,7 +1,7 @@
 Webots Project File version R2021a
-perspectives: 000000ff00000000fd00000003000000000000000000000000fc0100000001fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000000000140000008700ffffff000000010000011c00000224fc0200000001fb0000001400540065007800740045006400690074006f00720100000000000002240000004500ffffff00000003000005a0000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000005a00000008700ffffff000004820000022400000001000000020000000100000008fc00000000
+perspectives: 000000ff00000000fd00000003000000000000000000000000fc0100000001fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000ffffffff0000008700ffffff00000001000002020000022afc0200000001fb0000001400540065007800740045006400690074006f007201000000000000022a0000004500ffffff00000003000005a0000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000005a00000008700ffffff0000039c0000022a00000001000000020000000100000008fc00000000
 simulationViewPerspectives: 000000ff0000000100000002000001140000036e0100000002010000000100
-sceneTreePerspectives: 000000ff0000000100000002000000f5000000fa0100000002010000000200
+sceneTreePerspectives: 000000ff0000000100000002000000f8000000fa0100000002010000000200
 maximizedDockId: -1
 centralWidgetVisible: 1
 orthographicViewHeight: 1
